--- lib-multisrc/po2scans/build.gradle.kts
@@ -1,9 +1,0 @@
-plugins {
-    id("lib-multisrc")
-}
-
-baseVersionCode = 1
-
-dependencies {
-    api(project(":lib:dataimage"))
-}
