+++ src/tr/sadscans/src/eu/kanade/tachiyomi/extension/tr/sadscans/SadscansUrlActivity.kt
@@ -1,4 +1,4 @@
-package eu.kanade.tachiyomi.multisrc.po2scans
+package eu.kanade.tachiyomi.extension.tr.sadscans
 
 import android.app.Activity
 import android.content.ActivityNotFoundException
@@ -7,7 +7,7 @@ import android.os.Bundle
 import android.util.Log
 import kotlin.system.exitProcess
 
-class PO2ScansUrlActivity : Activity() {
+class SadscansUrlActivity : Activity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         val pathSegments = intent?.data?.pathSegments
@@ -15,17 +15,17 @@ class PO2ScansUrlActivity : Activity() {
             val slug = pathSegments[1]
             val mainIntent = Intent().apply {
                 action = "eu.kanade.tachiyomi.SEARCH"
-                putExtra("query", "${PO2Scans.SLUG_SEARCH_PREFIX}$slug")
+                putExtra("query", "${Sadscans.SLUG_SEARCH_PREFIX}$slug")
                 putExtra("filter", packageName)
             }
 
             try {
                 startActivity(mainIntent)
             } catch (e: ActivityNotFoundException) {
-                Log.e("PO2ScansUrlActivity", "Could not start activity", e)
+                Log.e("SadscansUrlActivity", "Could not start activity", e)
             }
         } else {
-            Log.e("PO2ScansUrlActivity", "could not parse URI from intent $intent")
+            Log.e("SadscansUrlActivity", "could not parse URI from intent $intent")
         }
 
         finish()
